+++ tools/lib/aas-v1/versions.js
@@ -4,5 +4,5 @@ module.exports = Object.freeze({
   protocolVersion: "2025-06-18",
   coreVersion: "1.0.0",
   metadataSchemaVersion: "1.0.0",
-  scorerVersion: "1.0.0",
+  scorerVersion: "1.1.0",
 });